fix: add conditional types export to manifest
Projects running TypeScript with moduleResolution set to `Node16`, `NodeNext` or `Bundler` will read the information in `package.json#exports` to determine the location of the types.

**The information in top-level `types` is discarded.**

The way TS resolves the types is shown here:
https://www.typescriptlang.org/docs/handbook/modules/reference.html#file-extension-substitution

The document above says that with the file extension substitution the types for the CJS version (`exports.require`) are searched for in `./dist/index.d.cts` and the types for the ESM (`exports.import`) are searched for in `./dist/index.esm.d.ts`, but none of these files exist. Instead, you have a single `./dist/index.d.ts` file.

To make TS look fot the types in there, you have to add the `"types": "./dist/index.d.ts"` entry as the first entry in `exports`.